pyproject.toml
[project]
name = "langgraph-agent"
version = "0.1.0"
description = "LangGraph agent with DDD and interface-agnostic adapters"
authors = [
  { name = "Felipe Mestre" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=0.0.40",
  "langchain>=0.1.0",
  "langchain-openai>=0.0.5",
  "langchain-community>=0.0.20",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "pymysql>=1.1.0",
  "requests>=2.31.0",
  "pydantic>=2.5.0",
  "python-dotenv>=1.0.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "streamlit>=1.28.0",
  "rich>=13.0.0",
  "pytest>=7.4.0"
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
